Add error handling for submodule download

+++ build/proxmox/build.sh
@@ -73,9 +73,14 @@ cd pve-kernel
 #  bypasses the process safely.
 # This curl skips certificate validation because Proxmox GIT WebUI doesn't send Let's Encrypt intermediate cert
 echo "Step 2.2: Downloading base kernel"
-curl -k "https://git.proxmox.com/?p=mirror_ubuntu-focal-kernel.git;a=snapshot;h=$(git submodule status submodules/ubuntu-focal | cut -c 2-41);sf=tgz" --output kernel.tgz
-tar -xf kernel.tgz -C submodules/ubuntu-focal/ --strip 1
-rm kernel.tgz
+curl -f -k "https://git.proxmox.com/?p=mirror_ubuntu-focal-kernel.git;a=snapshot;h=$(git submodule status submodules/ubuntu-focal | cut -c 2-41);sf=tgz" --output kernel.tgz || true
+
+if [[ -f "kernel.tgz" ]]; then
+  tar -xf kernel.tgz -C submodules/ubuntu-focal/ --strip 1
+  rm kernel.tgz
+else
+  echo "[-] Failed to download flat base kernel (will use git instead)"
+fi
